requirements.txt
# requisitos: instalacao dos pacotes abaixo
# pip install -r requirements.txt

pandas
mysql-connector-python
pillow
python-magic
flask
flask-login
mergedeep
libmagic

# for windows ;o)
python-magic-bin ; sys_platform == 'win32'
